refactor: Separate localStorage logic into utils and define WeatherData type
