+++ src/lib/types.ts
@@ -1,0 +1,9 @@
+export interface WeatherData {
+  dateSol: string;
+  dateUTC: string;
+  highCelsius: string;
+  lowCelsius: string;
+  pressure?: string;
+  sunrise?: string;
+  sunset?: string;
+}
